backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
distro==1.9.0
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2.2
django-allauth==65.9.0
django-appconf==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-fsm==3.0.0
django-fsm-log==3.1.0
django-subscriptions==2.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
oauthlib==3.2.2
openai==1.84.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0

# Celery & Background Processing
celery==5.4.0
redis==5.2.1
django-celery-beat==2.8.0
django-celery-results==2.5.1

# Redis Cache & Performance
django-redis==5.4.0
hiredis==3.1.0
